requirements.txt
fastapi~=0.115.5
pydantic~=2.9.2
python-dotenv~=1.0.1
pytest~=8.3.3
httpx~=0.27.2
uvicorn~=0.32.0
redis~=5.2.0